Changed operator "=" to "<=" for continue statement. This has two effects; a) It reduces time complexity and execution time by more or less half. On a input set of 1736 issues, 3011960 comparisons were made (O(n²)) over 2 minutes and 20.596 seconds.
After the change it was reduced to 1505980 operations (O(n²/2)) over 1 minute and 12.853 seconds.

b) The amount of identified duplicates (with threshold value 0.3)
increased (!) from 360 to 365. So when less comparisons were done... the
number of identified duplicates increased. I do not know why this
happens, and further investigation should be done. Even more peculiar,
when "<=" was replace with ">=" the number of identified duplicates was
only 360, even tough the exact same amount of operations were performed.

So the following happens when a certain separator is used;
	">=" --> 360 duplicates identified
	"==" --> 362 duplicates identified
	"<=" --> 365 duplicates identified